[refactor] split RegisterForm component
